refactor: move and rename SubjectListing, SubjectIcon and SubjectCardDouble
