+++ src/components/SharedComponents/SubjectIcon/SubjectIcon.stories.tsx
