+++ src/components/SharedComponents/SubjectIcon/index.ts
