--- src/components/SubjectCardDouble/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./SubjectCardDouble";
